refactor: adjustments to file paths
