+++ frontend/app/hooks/theme-service.js
